+++ src/Phan/AST/ContextNode.php
@@ -145,13 +145,20 @@ class ContextNode
          * @param Node|int|string|float|null $name_node
          * @throws FQSENException
          */
-        return \array_map(function ($name_node) : FullyQualifiedClassName {
-            return (new ContextNode(
+        $result = [];
+        foreach ($this->node->children as $name_node) {
+            $trait_fqsen = (new ContextNode(
                 $this->code_base,
                 $this->context,
                 $name_node
             ))->getTraitFQSEN([]);
-        }, $this->node->children);
+            if ($trait_fqsen) {
+                // Should never be null but check anyway
+                // TODO warn
+                $result[] = $trait_fqsen;
+            }
+        }
+        return $result;
     }
 
     /**
@@ -1072,6 +1079,7 @@ class ContextNode
                 return new Variable(
                     $this->context,
                     $variable_name,
+                    // @phan-suppress-next-line PhanTypeMismatchArgumentNullable
                     Variable::getUnionTypeOfHardcodedGlobalVariableWithName($variable_name),
                     0
                 );
@@ -1128,6 +1136,7 @@ class ContextNode
                     return new Variable(
                         $this->context,
                         $variable_name,
+                        // @phan-suppress-next-line PhanTypeMismatchArgumentNullable
                         Variable::getUnionTypeOfHardcodedGlobalVariableWithName($variable_name),
                         0
                     );