+++ src/Phan/Analysis/AssignOperatorAnalysisVisitor.php
@@ -221,11 +221,13 @@ class AssignOperatorAnalysisVisitor extends FlagVisitorImplementation
                 !$left->genericArrayElementTypes()->isEmpty()
                 && $left->nonArrayTypes()->isEmpty()
             ) || $left->isType($array_type);
+            // @phan-suppress-previous-line PhanTypeMismatchArgumentNullable false positive from static init
 
             $right_is_array = (
                 !$right->genericArrayElementTypes()->isEmpty()
                 && $right->nonArrayTypes()->isEmpty()
             ) || $right->isType($array_type);
+            // @phan-suppress-previous-line PhanTypeMismatchArgumentNullable false positive from static init
 
             if ($left_is_array || $right_is_array) {
                 if ($left_is_array && $right_is_array) {