+++ src/Phan/Language/Scope.php
@@ -113,7 +113,7 @@ abstract class Scope
     /**
      * @return FQSEN in which this scope was declared
      * (e.g. a FullyQualifiedFunctionName, FullyQualifiedClassName, etc.)
-     * @suppress PhanPossiblyNullTypeReturn callers should call hasFQSEN
+     * @suppress PhanPossiblyNullTypeReturn, PhanPartialTypeMismatchReturn callers should call hasFQSEN
      */
     public function getFQSEN()
     {