+++ src/Phan/LanguageServer/Protocol/Range.php
@@ -27,6 +27,7 @@ class Range
      */
     public $end;
 
+    /** @suppress PhanTypeMismatchProperty anything useful will be non-null */
     public function __construct(Position $start = null, Position $end = null)
     {
         $this->start = $start;